Del opp TidslinjeKombinert i komponenter og flytt sortert til soknadSortering
