+++ src/utils/gruppering.ts
@@ -2,8 +2,9 @@ import dayjs from 'dayjs'
 import { logger } from '@navikt/next-logger'
 
 import { KlippetSykepengesoknadRecord, Soknad } from '../queryhooks/useSoknader'
-import { Sortering } from '../components/soknad/ValgtSortering'
 import { Filter } from '../components/Filter'
+
+import { sortert } from './soknadSortering'
 
 import { Klipp, maxTom, minFom, perioderSomMangler, sykmeldingDager, sykmeldingOverlappendeDager } from './overlapp'
 import { passerAlleFilter } from './filterlogikk'
@@ -187,68 +188,6 @@ function grupperPaArbeidsgiver(gruppertPaSykmelding: Map<string, SykmeldingGrupp
     })
 
     return gruppering
-}
-
-export function sortert(a: [string, SykmeldingGruppering], b: [string, SykmeldingGruppering], sortering: Sortering) {
-    const aSykmeldingId = a.at(0) as string
-    const aSykmeldingGruppering = a.at(1) as SykmeldingGruppering
-    const bSykmeldingId = b.at(0) as string
-    const bSykmeldingGruppering = b.at(1) as SykmeldingGruppering
-
-    function mapTilSoknadProp(syk: SykmeldingGruppering, prop: string) {
-        return Array.from(syk.soknader.values()).map((sok) => (sok.soknad as unknown as Record<string, unknown>)[prop])
-    }
-
-    function mapTilKlippProp(syk: SykmeldingGruppering, prop: string) {
-        return Array.from(syk.soknader.values())
-            .flatMap((sok) => sok.klippingAvSoknad)
-            .concat(syk.klippingAvSykmelding)
-            .map((klipp) => (klipp as unknown as Record<string, unknown>)[prop])
-    }
-
-    function isGreater(a: unknown, b: unknown): boolean {
-        if (typeof a === 'string' && typeof b === 'string') return a > b
-        if (dayjs.isDayjs(a) && dayjs.isDayjs(b)) return a.valueOf() > b.valueOf()
-        return false
-    }
-
-    function maximum(previousValue: unknown, currentValue: unknown) {
-        if (isGreater(currentValue, previousValue)) return currentValue
-        else return previousValue
-    }
-
-    const tilDayjsVerdi = (verdi: unknown) => {
-        if (dayjs.isDayjs(verdi)) return verdi
-        if (typeof verdi === 'string') return dayjs(verdi)
-        return dayjs(0)
-    }
-
-    switch (sortering) {
-        case 'sykmelding skrevet': {
-            const verdiA = mapTilSoknadProp(aSykmeldingGruppering, 'sykmeldingUtskrevet').reduce(maximum, dayjs(0))
-            const verdiB = mapTilSoknadProp(bSykmeldingGruppering, 'sykmeldingUtskrevet').reduce(maximum, dayjs(0))
-            return isGreater(verdiA, verdiB) ? -1 : 1
-        }
-        case 'opprettet': {
-            const verdiA = aSykmeldingId.endsWith('_GHOST')
-                ? mapTilKlippProp(aSykmeldingGruppering, 'timestamp').reduce(maximum, dayjs(0))
-                : mapTilSoknadProp(aSykmeldingGruppering, 'opprettetDato').reduce(maximum, dayjs(0))
-            const verdiB = bSykmeldingId.endsWith('_GHOST')
-                ? mapTilKlippProp(bSykmeldingGruppering, 'timestamp').reduce(maximum, dayjs(0))
-                : mapTilSoknadProp(bSykmeldingGruppering, 'opprettetDato').reduce(maximum, dayjs(0))
-            return isGreater(verdiA, verdiB) ? -1 : 1
-        }
-        default:
-        case 'tom': {
-            const verdiA = aSykmeldingId.endsWith('_GHOST')
-                ? mapTilKlippProp(aSykmeldingGruppering, 'tom').map(tilDayjsVerdi).reduce(maximum, dayjs('2000-01-01'))
-                : mapTilSoknadProp(aSykmeldingGruppering, 'tom').map(tilDayjsVerdi).reduce(maximum, dayjs('2000-01-01'))
-            const verdiB = bSykmeldingId.endsWith('_GHOST')
-                ? mapTilKlippProp(bSykmeldingGruppering, 'tom').map(tilDayjsVerdi).reduce(maximum, dayjs('2000-01-01'))
-                : mapTilSoknadProp(bSykmeldingGruppering, 'tom').map(tilDayjsVerdi).reduce(maximum, dayjs('2000-01-01'))
-            return isGreater(verdiA, verdiB) ? -1 : 1
-        }
-    }
 }
 
 function leggTilOppholdUtland(soknader: Soknad[], filter: Filter[], gruppering: Map<string, ArbeidsgiverGruppering>) {
